Moved ClientVersion and ClientLanguage to Network project
We need it in the chat server, and there we don't want to add a reference to our data model.
