+++ src/GameServer/MessageHandler/CharacterMoveHandlerPlugIn.cs
@@ -5,7 +5,7 @@
 namespace MUnique.OpenMU.GameServer.MessageHandler
 {
     using System.Runtime.InteropServices;
-    using MUnique.OpenMU.DataModel.Configuration;
+    using MUnique.OpenMU.Network.PlugIns;
     using MUnique.OpenMU.PlugIns;
 
     /// <summary>
