+++ src/GameServer/RemoteView/Login/ShowLoginWindowPlugIn.cs
@@ -6,10 +6,10 @@ namespace MUnique.OpenMU.GameServer.RemoteView.Login
 {
     using System;
     using System.Runtime.InteropServices;
-    using MUnique.OpenMU.DataModel.Configuration;
     using MUnique.OpenMU.GameLogic.Views;
     using MUnique.OpenMU.GameLogic.Views.Login;
     using MUnique.OpenMU.Network;
+    using MUnique.OpenMU.Network.PlugIns;
     using MUnique.OpenMU.PlugIns;
 
     /// <summary>
