+++ src/GameServer/RemoteView/RemotePlayer.cs
@@ -7,11 +7,11 @@ namespace MUnique.OpenMU.GameServer.RemoteView
     using System;
     using System.Buffers;
     using log4net;
-    using MUnique.OpenMU.DataModel.Configuration;
     using MUnique.OpenMU.GameLogic;
     using MUnique.OpenMU.GameLogic.Views;
     using MUnique.OpenMU.GameServer.MessageHandler;
     using MUnique.OpenMU.Network;
+    using MUnique.OpenMU.Network.PlugIns;
     using MUnique.OpenMU.PlugIns;
 
     /// <summary>
