+++ src/Network/PlugIns/ClientLanguage.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace MUnique.OpenMU.DataModel.Configuration
+namespace MUnique.OpenMU.Network.PlugIns
 {
     /// <summary>
     /// The language of the client.
